Added previewFrame. Fixes #1, needs improvement.
I don't think ffmpeg is the right choice for this. It's hard to say.
I might hotwire AviSynth into this process as well.
Should be faster with ffindex being there.

This at least lays the groundwork for more issues.
Also, tabPage1 is just a temporary thing to show off the preview.
Need to figure out where to actually put it in the real design.

CropForm is removed for now.