Added initial SubtitleFilter
Although internal subs are listed as an option, they don't work at the
moment.
I need a way to extract tracks from the input files, for internal
subtitles and attached fonts.
Luckily, assrender can use attached fonts if they're extracted, so that
shouldn't prove much of a challenge when it's ready.